+++ CHANGELOG.md
@@ -27,6 +27,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Made `Trompt` output 2-dim embeddings in `forward`
+
 ### Deprecated
 
 ### Removed